Move class under Tests
